+++ tests/test_player.py
@@ -68,6 +68,18 @@ class TestPlayer(unittest.TestCase):
         player.forfeit_tiles([Tile('B', 3), Tile('C', 1)])
         self.assertEqual(player.tiles, [Tile('A', 1)])
 
+    def test_set_name(self):
+        player = Player()
+        player.set_name('Dino')
+        self.assertEqual(player.get_name(), 'Dino')
+
+    def test_get_score(self):
+        player = Player()
+        player.increase_score(2)
+        self.assertEqual(player.get_score(), 2)
+
+
+
 
 if __name__ == '__main__':
     unittest.main()